Updating a CMakeLists.txt file and changing comments in CFML_Propagk.f90
git-svn-id: https://forge.ill.fr/svn/crysfml@804 06508225-4b59-0410-a2a9-fdab55894d57

+++ Src/CFML_Propagk.f90
@@ -286,9 +286,9 @@
 
     !!----
     !!---- Subroutine K_Star(K,Spacegroup,Gk)
-    !!----    integer, dimension(3),   intent(in)  :: k           !  In ->
-    !!----    type (Space_Group_Type), intent(in)  :: SpaceGroup  !  In ->
-    !!----    Type (Group_k_Type),     intent(out) :: Groupk      ! Out ->
+    !!----    real(kind=cp),dimension(3), intent(in)  :: k           !  In ->
+    !!----    type (Space_Group_Type),    intent(in)  :: SpaceGroup  !  In ->
+    !!----    Type (Group_k_Type),        intent(out) :: Groupk      ! Out ->
     !!----
     !!----    Calculate the star of the propagation vector
     !!----    and the group of the vector k. Construct the object
